@@ -10,38 +10,46 @@
     "lint": "next lint"
   },
   "dependencies": {
-    "@ai-sdk/anthropic": "^0.0.51",
-    "@ai-sdk/azure": "^0.0.41",
-    "@ai-sdk/google": "^0.0.51",
-    "@ai-sdk/openai": "^0.0.63",
+    "@ai-sdk/anthropic": "^1.1.5",
+    "@ai-sdk/azure": "^1.1.5",
+    "@ai-sdk/deepseek": "^0.1.6",
+    "@ai-sdk/fireworks": "^0.1.6",
+    "@ai-sdk/google": "^1.1.5",
+    "@ai-sdk/groq": "^1.1.6",
+    "@ai-sdk/openai": "^1.1.5",
+    "@ai-sdk/xai": "^1.1.10",
     "@radix-ui/react-alert-dialog": "^1.0.5",
     "@radix-ui/react-avatar": "^1.0.4",
     "@radix-ui/react-checkbox": "^1.0.4",
     "@radix-ui/react-collapsible": "^1.0.3",
-    "@radix-ui/react-dialog": "^1.0.5",
+    "@radix-ui/react-dialog": "^1.1.5",
     "@radix-ui/react-dropdown-menu": "^2.0.6",
     "@radix-ui/react-label": "^2.0.2",
+    "@radix-ui/react-popover": "^1.1.5",
     "@radix-ui/react-select": "^2.1.2",
     "@radix-ui/react-separator": "^1.0.3",
     "@radix-ui/react-slider": "^1.1.2",
     "@radix-ui/react-slot": "^1.0.2",
     "@radix-ui/react-switch": "^1.0.3",
+    "@radix-ui/react-toggle": "^1.1.1",
     "@radix-ui/react-tooltip": "^1.0.7",
     "@tailwindcss/typography": "^0.5.12",
     "@upstash/redis": "^1.34.0",
-    "ai": "^3.4.7",
+    "ai": "^4.1.9",
     "class-variance-authority": "^0.7.0",
     "clsx": "^2.1.0",
+    "cmdk": "1.0.0",
     "embla-carousel-react": "^8.0.0",
     "exa-js": "^1.0.12",
     "jsdom": "^22.1.0",
     "katex": "^0.16.10",
-    "lucide-react": "^0.363.0",
-    "next": "^14.2.3",
+    "lucide-react": "^0.474.0",
+    "next": "^15.2.0",
     "next-themes": "^0.3.0",
     "node-html-parser": "^6.1.13",
-    "ollama-ai-provider": "^0.15.1",
-    "react-dom": "^18",
+    "ollama-ai-provider": "^1.2.0",
+    "react": "^19.0.0",
+    "react-dom": "^19.0.0",
     "react-icons": "^5.0.1",
     "react-markdown": "^8.0.7",
     "react-syntax-highlighter": "^15.5.0",
@@ -52,7 +60,7 @@
     "remark-gfm": "^3.0.1",
     "remark-math": "^5.1.1",
     "sonner": "^1.4.41",
-    "tailwind-merge": "^2.2.2",
+    "tailwind-merge": "^2.6.0",
     "tailwindcss-animate": "^1.0.7",
     "zod": "^3.23.8"
   },
