@@ -1,14 +1,9 @@
 import { Chat } from '@/components/chat'
+import { getModels } from '@/lib/config/models'
 import { generateId } from 'ai'
-import { AI } from './actions'
 
-export const maxDuration = 60
-
-export default function Page() {
+export default async function Page() {
   const id = generateId()
-  return (
-    <AI initialAIState={{ chatId: id, messages: [] }}>
-      <Chat id={id} />
-    </AI>
-  )
+  const models = await getModels()
+  return <Chat id={id} models={models} />
 }
