@@ -1,26 +1,30 @@
-import { notFound, redirect } from 'next/navigation'
 import { Chat } from '@/components/chat'
 import { getChat } from '@/lib/actions/chat'
-import { AI } from '@/app/actions'
+import { getModels } from '@/lib/config/models'
+import { convertToUIMessages } from '@/lib/utils'
+import { notFound, redirect } from 'next/navigation'
 
 export const maxDuration = 60
 
-export interface SearchPageProps {
-  params: {
-    id: string
-  }
-}
-
-export async function generateMetadata({ params }: SearchPageProps) {
-  const chat = await getChat(params.id, 'anonymous')
+export async function generateMetadata(props: {
+  params: Promise<{ id: string }>
+}) {
+  const { id } = await props.params
+  const chat = await getChat(id, 'anonymous')
   return {
     title: chat?.title.toString().slice(0, 50) || 'Search'
   }
 }
 
-export default async function SearchPage({ params }: SearchPageProps) {
+export default async function SearchPage(props: {
+  params: Promise<{ id: string }>
+}) {
   const userId = 'anonymous'
-  const chat = await getChat(params.id, userId)
+  const { id } = await props.params
+
+  const chat = await getChat(id, userId)
+  // convertToUIMessages for useChat hook
+  const messages = convertToUIMessages(chat?.messages || [])
 
   if (!chat) {
     redirect('/')
@@ -30,14 +34,6 @@ export default async function SearchPage({ params }: SearchPageProps) {
     notFound()
   }
 
-  return (
-    <AI
-      initialAIState={{
-        chatId: chat.id,
-        messages: chat.messages
-      }}
-    >
-      <Chat id={params.id} />
-    </AI>
-  )
+  const models = await getModels()
+  return <Chat id={id} savedMessages={messages} models={models} />
 }
