@@ -1,23 +1,19 @@
 import { Chat } from '@/components/chat'
+import { getModels } from '@/lib/config/models'
 import { generateId } from 'ai'
-import { AI } from '@/app/actions'
 import { redirect } from 'next/navigation'
 
 export const maxDuration = 60
 
-export default function Page({
-  searchParams
-}: {
-  searchParams: { q: string }
+export default async function SearchPage(props: {
+  searchParams: Promise<{ q: string }>
 }) {
-  if (!searchParams.q) {
+  const { q } = await props.searchParams
+  if (!q) {
     redirect('/')
   }
-  const id = generateId()
 
-  return (
-    <AI initialAIState={{ chatId: id, messages: [] }}>
-      <Chat id={id} query={searchParams.q} />
-    </AI>
-  )
+  const id = generateId()
+  const models = await getModels()
+  return <Chat id={id} query={q} models={models} />
 }
