@@ -1,37 +1,54 @@
 'use client'
 
-import { Section } from './section'
-import { StreamableValue, useStreamableValue } from 'ai/rsc'
-import { BotMessage } from './message'
-import { useEffect, useState } from 'react'
+import { Text } from 'lucide-react'
+import { CollapsibleMessage } from './collapsible-message'
 import { DefaultSkeleton } from './default-skeleton'
+import { BotMessage } from './message'
+import { MessageActions } from './message-actions'
 
 export type AnswerSectionProps = {
-  result?: StreamableValue<string>
-  hasHeader?: boolean
+  content: string
+  isOpen: boolean
+  onOpenChange: (open: boolean) => void
+  chatId?: string
 }
 
 export function AnswerSection({
-  result,
-  hasHeader = true
+  content,
+  isOpen,
+  onOpenChange,
+  chatId
 }: AnswerSectionProps) {
-  const [data, error, pending] = useStreamableValue(result)
-  const [content, setContent] = useState<string>('')
+  const enableShare = process.env.NEXT_PUBLIC_ENABLE_SHARE === 'true'
 
-  useEffect(() => {
-    if (!data) return
-    setContent(data)
-  }, [data])
-
-  return (
-    <div>
-      {content.length > 0 ? (
-        <Section title={hasHeader ? 'Answer' : undefined}>
-          <BotMessage content={content} />
-        </Section>
-      ) : (
-        <DefaultSkeleton />
-      )}
+  const header = (
+    <div className="flex items-center gap-1">
+      <Text size={16} />
+      <div>Answer</div>
     </div>
+  )
+  const message = content ? (
+    <div className="flex flex-col gap-1">
+      <BotMessage message={content} />
+      <MessageActions
+        message={content}
+        chatId={chatId}
+        enableShare={enableShare}
+      />
+    </div>
+  ) : (
+    <DefaultSkeleton />
+  )
+  return (
+    <CollapsibleMessage
+      role="assistant"
+      isCollapsible={false}
+      header={header}
+      isOpen={isOpen}
+      onOpenChange={onOpenChange}
+      showBorder={false}
+    >
+      {message}
+    </CollapsibleMessage>
   )
 }
