@@ -1,20 +1,21 @@
 'use client'
 
-import { useState, useTransition } from 'react'
-import { Button } from './ui/button'
+import { shareChat } from '@/lib/actions/chat'
+import { useCopyToClipboard } from '@/lib/hooks/use-copy-to-clipboard'
+import { cn } from '@/lib/utils'
 import { Share } from 'lucide-react'
+import { useState, useTransition } from 'react'
+import { toast } from 'sonner'
+import { Button } from './ui/button'
 import {
   Dialog,
   DialogContent,
+  DialogDescription,
   DialogFooter,
   DialogHeader,
-  DialogTrigger,
-  DialogDescription,
-  DialogTitle
+  DialogTitle,
+  DialogTrigger
 } from './ui/dialog'
-import { shareChat } from '@/lib/actions/chat'
-import { toast } from 'sonner'
-import { useCopyToClipboard } from '@/lib/hooks/use-copy-to-clipboard'
 import { Spinner } from './ui/spinner'
 
 interface ChatShareProps {
@@ -67,7 +68,7 @@ export function ChatShare({ chatId, className }: ChatShareProps) {
       >
         <DialogTrigger asChild>
           <Button
-            className="rounded-full"
+            className={cn('rounded-full')}
             size="icon"
             variant={'ghost'}
             onClick={() => setOpen(true)}
