@@ -1,30 +1,88 @@
 'use client'
 
+import { CHAT_ID } from '@/lib/constants'
+import { Model } from '@/lib/types/models'
+import { Message, useChat } from 'ai/react'
 import { useEffect } from 'react'
-import { usePathname } from 'next/navigation'
-import { ChatPanel } from './chat-panel'
+import { toast } from 'sonner'
 import { ChatMessages } from './chat-messages'
-import { useUIState } from 'ai/rsc'
+import { ChatPanel } from './chat-panel'
 
-type ChatProps = {
-  id?: string
+export function Chat({
+  id,
+  savedMessages = [],
+  query,
+  models
+}: {
+  id: string
+  savedMessages?: Message[]
   query?: string
-}
-
-export function Chat({ id, query }: ChatProps) {
-  const path = usePathname()
-  const [messages] = useUIState()
+  models?: Model[]
+}) {
+  const {
+    messages,
+    input,
+    handleInputChange,
+    handleSubmit,
+    isLoading,
+    setMessages,
+    stop,
+    append,
+    data,
+    setData
+  } = useChat({
+    initialMessages: savedMessages,
+    id: CHAT_ID,
+    body: {
+      id
+    },
+    onFinish: () => {
+      window.history.replaceState({}, '', `/search/${id}`)
+    },
+    onError: error => {
+      toast.error(`Error in chat: ${error.message}`)
+    },
+    sendExtraMessageFields: false // Disable extra message fields
+  })
 
   useEffect(() => {
-    if (!path.includes('search') && messages.length === 1) {
-      window.history.replaceState({}, '', `/search/${id}`)
-    }
-  }, [id, path, messages, query])
+    setMessages(savedMessages)
+  }, [id])
+
+  const onQuerySelect = (query: string) => {
+    append({
+      role: 'user',
+      content: query
+    })
+  }
+
+  const onSubmit = (e: React.FormEvent<HTMLFormElement>) => {
+    e.preventDefault()
+    setData(undefined) // reset data to clear tool call
+    handleSubmit(e)
+  }
 
   return (
-    <div className="px-8 sm:px-12 pt-12 md:pt-14 pb-14 md:pb-24 max-w-3xl mx-auto flex flex-col space-y-3 md:space-y-4">
-      <ChatMessages messages={messages} />
-      <ChatPanel messages={messages} query={query} />
+    <div className="flex flex-col w-full max-w-3xl pt-14 pb-60 mx-auto stretch">
+      <ChatMessages
+        messages={messages}
+        data={data}
+        onQuerySelect={onQuerySelect}
+        isLoading={isLoading}
+        chatId={id}
+      />
+      <ChatPanel
+        input={input}
+        handleInputChange={handleInputChange}
+        handleSubmit={onSubmit}
+        isLoading={isLoading}
+        messages={messages}
+        setMessages={setMessages}
+        stop={stop}
+        query={query}
+        append={append}
+        models={models}
+      />
     </div>
   )
 }
