@@ -1,75 +1,76 @@
-'use client'
-
-import React, { useEffect, useState } from 'react'
+import { cn } from '@/lib/utils'
+import { ChevronDown, UserCircle2 } from 'lucide-react'
 import {
   Collapsible,
-  CollapsibleTrigger,
-  CollapsibleContent
-} from '@radix-ui/react-collapsible'
-import { Button } from './ui/button'
-import { ChevronDown } from 'lucide-react'
-import { StreamableValue, useStreamableValue } from 'ai/rsc'
-import { cn } from '@/lib/utils'
+  CollapsibleContent,
+  CollapsibleTrigger
+} from './ui/collapsible'
+import { IconLogo } from './ui/icons'
 import { Separator } from './ui/separator'
 
 interface CollapsibleMessageProps {
-  message: {
-    id: string
-    isCollapsed?: StreamableValue<boolean>
-    component: React.ReactNode
-  }
-  isLastMessage?: boolean
+  children: React.ReactNode
+  role: 'user' | 'assistant'
+  isCollapsible?: boolean
+  isOpen?: boolean
+  header?: React.ReactNode
+  onOpenChange?: (open: boolean) => void
+  showBorder?: boolean
+  showIcon?: boolean
 }
 
-export const CollapsibleMessage: React.FC<CollapsibleMessageProps> = ({
-  message,
-  isLastMessage = false
-}) => {
-  const [data] = useStreamableValue(message.isCollapsed)
-  const isCollapsed = data ?? false
-  const [open, setOpen] = useState(isLastMessage)
-
-  useEffect(() => {
-    setOpen(isLastMessage)
-  }, [isCollapsed, isLastMessage])
-
-  // if not collapsed, return the component
-  if (!isCollapsed) {
-    return message.component
-  }
+export function CollapsibleMessage({
+  children,
+  role,
+  isCollapsible = false,
+  isOpen = true,
+  header,
+  onOpenChange,
+  showBorder = true,
+  showIcon = true
+}: CollapsibleMessageProps) {
+  const content = <div className="py-2 flex-1">{children}</div>
 
   return (
-    <Collapsible
-      open={open}
-      onOpenChange={value => {
-        setOpen(value)
-      }}
-    >
-      <CollapsibleTrigger asChild>
+    <div className="flex gap-3">
+      <div className="relative flex flex-col items-center">
+        <div className={cn('mt-[10px] w-5', role === 'assistant' && 'mt-4')}>
+          {showIcon &&
+            (role === 'user' ? (
+              <UserCircle2 size={20} className="text-muted-foreground" />
+            ) : (
+              <IconLogo className="size-5" />
+            ))}
+        </div>
+      </div>
+
+      {isCollapsible ? (
         <div
           className={cn(
-            'w-full flex justify-end',
-            !isCollapsed ? 'hidden' : ''
+            'flex-1 rounded-2xl p-4',
+            showBorder && 'border border-border/50'
           )}
         >
-          <Button
-            variant="ghost"
-            size={'icon'}
-            className={cn('-mt-3 rounded-full')}
+          <Collapsible
+            open={isOpen}
+            onOpenChange={onOpenChange}
+            className="w-full"
           >
-            <ChevronDown
-              size={14}
-              className={cn(
-                open ? 'rotate-180' : 'rotate-0',
-                'h-4 w-4 transition-all'
-              )}
-            />
-            <span className="sr-only">collapse</span>
-          </Button>
+            <CollapsibleTrigger className="flex items-center justify-between w-full group">
+              <div className="flex items-center justify-between w-full gap-2">
+                {header && <div className="text-sm w-full">{header}</div>}
+                <ChevronDown className="h-4 w-4 text-muted-foreground transition-transform duration-200 group-data-[state=open]:rotate-180" />
+              </div>
+            </CollapsibleTrigger>
+            <CollapsibleContent className="data-[state=closed]:animate-collapse-up data-[state=open]:animate-collapse-down">
+              <Separator className="my-4 border-border/50" />
+              {content}
+            </CollapsibleContent>
+          </Collapsible>
         </div>
-      </CollapsibleTrigger>
-      <CollapsibleContent>{message.component}</CollapsibleContent>
-      {!open && <Separator className="my-2 bg-muted" />}
-    </Collapsible>
+      ) : (
+        <div className="flex-1 rounded-2xl px-4">{content}</div>
+      )}
+    </div>
   )
 }
