@@ -1,13 +1,27 @@
 'use client'
 
-import React from 'react'
 import { Skeleton } from './ui/skeleton'
 
 export const DefaultSkeleton = () => {
   return (
-    <div className="flex flex-col gap-2 pb-4">
+    <div className="flex flex-col gap-2 pb-4 pt-2">
       <Skeleton className="h-6 w-48" />
       <Skeleton className="w-full h-6" />
+    </div>
+  )
+}
+
+export function SearchSkeleton() {
+  return (
+    <div className="flex flex-wrap gap-2 pb-0.5">
+      {[...Array(4)].map((_, index) => (
+        <div
+          key={index}
+          className="w-[calc(50%-0.5rem)] md:w-[calc(25%-0.5rem)]"
+        >
+          <Skeleton className="h-20 w-full" />
+        </div>
+      ))}
     </div>
   )
 }
