@@ -1,12 +1,12 @@
+import { cn } from '@/lib/utils'
 import React from 'react'
+import HistoryContainer from './history-container'
 import { ModeToggle } from './mode-toggle'
 import { IconLogo } from './ui/icons'
-import { cn } from '@/lib/utils'
-import HistoryContainer from './history-container'
 
 export const Header: React.FC = async () => {
   return (
-    <header className="fixed w-full p-1 md:p-2 flex justify-between items-center z-10 backdrop-blur md:backdrop-blur-none bg-background/80 md:bg-transparent">
+    <header className="fixed w-full p-2 flex justify-between items-center z-10 backdrop-blur lg:backdrop-blur-none bg-background/80 lg:bg-transparent">
       <div>
         <a href="/">
           <IconLogo className={cn('w-5 h-5')} />
@@ -15,7 +15,7 @@ export const Header: React.FC = async () => {
       </div>
       <div className="flex gap-0.5">
         <ModeToggle />
-        <HistoryContainer location="header" />
+        <HistoryContainer />
       </div>
     </header>
   )
