@@ -2,18 +2,15 @@ import React from 'react'
 import { History } from './history'
 import { HistoryList } from './history-list'
 
-type HistoryContainerProps = {
-  location: 'sidebar' | 'header'
-}
+const HistoryContainer: React.FC = async () => {
+  const enableSaveChatHistory = process.env.ENABLE_SAVE_CHAT_HISTORY === 'true'
+  if (!enableSaveChatHistory) {
+    return null
+  }
 
-const HistoryContainer: React.FC<HistoryContainerProps> = async ({
-  location
-}) => {
   return (
-    <div
-      className={location === 'header' ? 'block sm:hidden' : 'hidden sm:block'}
-    >
-      <History location={location}>
+    <div>
+      <History>
         <HistoryList userId="anonymous" />
       </History>
     </div>
