@@ -1,7 +1,6 @@
 'use client'
 
-import { useTransition } from 'react'
-import { useRouter } from 'next/navigation'
+import { Button } from '@/components/ui/button'
 import {
   Sheet,
   SheetContent,
@@ -9,23 +8,18 @@ import {
   SheetTitle,
   SheetTrigger
 } from '@/components/ui/sheet'
-import { Button } from '@/components/ui/button'
-import { ChevronLeft, Menu } from 'lucide-react'
-import { cn } from '@/lib/utils'
-import { History as HistoryIcon } from 'lucide-react'
-import { Suspense } from 'react'
+import { History as HistoryIcon, Menu } from 'lucide-react'
+import { useRouter } from 'next/navigation'
+import { Suspense, useTransition } from 'react'
 import { HistorySkeleton } from './history-skeleton'
-import { useAppState } from '@/lib/utils/app-state'
 
 type HistoryProps = {
-  location: 'sidebar' | 'header'
   children?: React.ReactNode
 }
 
-export function History({ location, children }: HistoryProps) {
+export function History({ children }: HistoryProps) {
   const router = useRouter()
   const [isPending, startTransition] = useTransition()
-  const { isGenerating, setIsGenerating } = useAppState()
 
   const onOpenChange = (open: boolean) => {
     if (open) {
@@ -38,15 +32,8 @@ export function History({ location, children }: HistoryProps) {
   return (
     <Sheet onOpenChange={onOpenChange}>
       <SheetTrigger asChild>
-        <Button
-          variant="ghost"
-          size="icon"
-          className={cn({
-            'rounded-full text-foreground/30': location === 'sidebar'
-          })}
-          disabled={isGenerating}
-        >
-          {location === 'header' ? <Menu /> : <ChevronLeft size={16} />}
+        <Button variant="ghost" size="icon">
+          <Menu />
         </Button>
       </SheetTrigger>
       <SheetContent className="w-64 rounded-tl-xl rounded-bl-xl">
