@@ -1,46 +1,70 @@
 'use client'
 
-import { SearchResults } from './search-results'
-import { DefaultSkeleton } from './default-skeleton'
-import { SearchResultsImageSection } from './search-results-image'
-import { Section } from './section'
-import { ToolBadge } from './tool-badge'
+import { CHAT_ID } from '@/lib/constants'
 import type { SearchResults as TypeSearchResults } from '@/lib/types'
-import { StreamableValue, useStreamableValue } from 'ai/rsc'
+import { ToolInvocation } from 'ai'
+import { useChat } from 'ai/react'
+import { CollapsibleMessage } from './collapsible-message'
+import { SearchSkeleton } from './default-skeleton'
+import { SearchResults } from './search-results'
+import { SearchResultsImageSection } from './search-results-image'
+import { Section, ToolArgsSection } from './section'
 
-export type SearchSectionProps = {
-  result?: StreamableValue<string>
-  includeDomains?: string[]
+interface SearchSectionProps {
+  tool: ToolInvocation
+  isOpen: boolean
+  onOpenChange: (open: boolean) => void
 }
 
-export function SearchSection({ result, includeDomains }: SearchSectionProps) {
-  const [data, error, pending] = useStreamableValue(result)
-  const searchResults: TypeSearchResults = data ? JSON.parse(data) : undefined
+export function SearchSection({
+  tool,
+  isOpen,
+  onOpenChange
+}: SearchSectionProps) {
+  const { isLoading } = useChat({
+    id: CHAT_ID
+  })
+  const isToolLoading = tool.state === 'call'
+  const searchResults: TypeSearchResults =
+    tool.state === 'result' ? tool.result : undefined
+  const query = tool.args?.query as string | undefined
+  const includeDomains = tool.args?.includeDomains as string[] | undefined
   const includeDomainsString = includeDomains
     ? ` [${includeDomains.join(', ')}]`
     : ''
+
+  const header = (
+    <ToolArgsSection
+      tool="search"
+      number={searchResults?.results?.length}
+    >{`${query}${includeDomainsString}`}</ToolArgsSection>
+  )
+
   return (
-    <div>
-      {!pending && data ? (
-        <>
-          <Section size="sm" className="pt-2 pb-0">
-            <ToolBadge tool="search">{`${searchResults.query}${includeDomainsString}`}</ToolBadge>
+    <CollapsibleMessage
+      role="assistant"
+      isCollapsible={true}
+      header={header}
+      isOpen={isOpen}
+      onOpenChange={onOpenChange}
+    >
+      {searchResults &&
+        searchResults.images &&
+        searchResults.images.length > 0 && (
+          <Section>
+            <SearchResultsImageSection
+              images={searchResults.images}
+              query={query}
+            />
           </Section>
-          {searchResults.images && searchResults.images.length > 0 && (
-            <Section title="Images">
-              <SearchResultsImageSection
-                images={searchResults.images}
-                query={searchResults.query}
-              />
-            </Section>
-          )}
-          <Section title="Sources">
-            <SearchResults results={searchResults.results} />
-          </Section>
-        </>
-      ) : (
-        <DefaultSkeleton />
-      )}
-    </div>
+        )}
+      {isLoading && isToolLoading ? (
+        <SearchSkeleton />
+      ) : searchResults?.results ? (
+        <Section title="Sources">
+          <SearchResults results={searchResults.results} />
+        </Section>
+      ) : null}
+    </CollapsibleMessage>
   )
 }
