@@ -3,6 +3,7 @@
 import { cn } from '@/lib/utils'
 import {
   BookCheck,
+  Check,
   Film,
   Image,
   MessageCircleMore,
@@ -11,7 +12,10 @@ import {
   Search
 } from 'lucide-react'
 import React from 'react'
+import { ToolBadge } from './tool-badge'
+import { Badge } from './ui/badge'
 import { Separator } from './ui/separator'
+import { StatusIndicator } from './ui/status-indicator'
 
 type SectionProps = {
   children: React.ReactNode
@@ -60,18 +64,42 @@ export const Section: React.FC<SectionProps> = ({
       {separator && <Separator className="my-2 bg-primary/10" />}
       <section
         className={cn(
-          ` ${size === 'sm' ? 'py-1' : size === 'lg' ? 'py-4' : 'py-2'}`,
+          ` ${size === 'sm' ? 'py-1' : size === 'lg' ? 'py-4' : 'py-1'}`,
           className
         )}
       >
         {title && (
-          <h2 className="flex items-center leading-none py-2">
+          <Badge
+            variant="secondary"
+            className="flex items-center leading-none w-fit my-1"
+          >
             {icon}
             {title}
-          </h2>
+          </Badge>
         )}
         {children}
       </section>
     </>
+  )
+}
+
+export function ToolArgsSection({
+  children,
+  tool,
+  number
+}: {
+  children: React.ReactNode
+  tool: string
+  number?: number
+}) {
+  return (
+    <Section size="sm" className="py-0 flex items-center justify-between">
+      <ToolBadge tool={tool}>{children}</ToolBadge>
+      {number && (
+        <StatusIndicator icon={Check} iconClassName="text-green-500">
+          {number} results
+        </StatusIndicator>
+      )}
+    </Section>
   )
 }
