@@ -1,5 +1,5 @@
 import React from 'react'
-import { Search } from 'lucide-react'
+import { Link, Search, Video } from 'lucide-react'
 import { Badge } from './ui/badge'
 
 type ToolBadgeProps = {
@@ -14,7 +14,9 @@ export const ToolBadge: React.FC<ToolBadgeProps> = ({
   className
 }) => {
   const icon: Record<string, React.ReactNode> = {
-    search: <Search size={14} />
+    search: <Search size={14} />,
+    retrieve: <Link size={14} />,
+    video_search: <Video size={14} />
   }
 
   return (
