@@ -3,9 +3,9 @@ import { z } from 'zod'
 
 export const searchSchema = z.object({
   query: z.string().describe('The query to search for'),
-  max_results: z.coerce
+  max_results: z
     .number()
-    .describe('The maximum number of results to return'),
+    .describe('The maximum number of results to return. default is 20'),
   search_depth: z
     .string()
     .describe(
@@ -13,13 +13,11 @@ export const searchSchema = z.object({
     ),
   include_domains: z
     .array(z.string())
-    .optional()
     .describe(
       'A list of domains to specifically include in the search results. Default is None, which includes all domains.'
     ),
   exclude_domains: z
     .array(z.string())
-    .optional()
     .describe(
       "A list of domains to specifically exclude from the search results. Default is None, which doesn't exclude any domains."
     )
